pyproject.toml
[project]
name            = "dr-kudale-s-clinic"
version         = "0.1.0"
description     = "Add your description here"
authors         = [{ name = "aspizu", email = "aspizu@protonmail.com" }]
dependencies    = ["python-telegram-bot>=21.0.1", "rich>=13.7.1"]
readme          = "README.md"
requires-python = ">= 3.8"

[tool.rye.scripts]
main = { call = "dr_kudale_s_clinic:main" }

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed          = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dr_kudale_s_clinic"]
